core_api/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.6.15
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fast-depends==2.4.12
fastapi==0.115.14
faststream==0.5.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
pamqp==3.3.0
propcache==0.3.2
pyasn1==0.6.1
pybreaker==1.3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-jose==3.5.0
pytz==2025.2
redis==6.2.0
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
watchfiles==1.1.0
yarl==1.20.1
